Make ShippingRateSection shared component and move FlatShippingRatesInputCards into it.
