+++ js/src/components/shipping-rate-section/index.js
@@ -1,0 +1,1 @@
+export { default } from './shipping-rate-section';
