+++ js/src/components/shipping-rate-section/minimum-order-card/minimum-order-card.scss
